Tidy fixes

+++ src/Client/BuzzHouse/Generator/SQLTable.cpp
@@ -1419,14 +1419,14 @@ void StatementGenerator::generateEngineDetails(
     }
     else if (te->has_engine() && b.isGenerateRandomEngine())
     {
-        te->add_params()->set_num(rg.nextInFullRange());
+        te->add_params()->set_num(static_cast<uint32_t>(rg.nextInFullRange()));
         if (rg.nextBool())
         {
             std::uniform_int_distribution<uint32_t> string_length_dist(0, fc.max_string_length);
             std::uniform_int_distribution<uint64_t> nested_rows_dist(fc.min_nested_rows, fc.max_nested_rows);
 
             te->add_params()->set_num(string_length_dist(rg.generator));
-            te->add_params()->set_num(nested_rows_dist(rg.generator));
+            te->add_params()->set_num(static_cast<uint32_t>(nested_rows_dist(rg.generator)));
         }
     }
     else if (te->has_engine() && b.isKeeperMapEngine())
@@ -1436,7 +1436,7 @@ void StatementGenerator::generateEngineDetails(
         {
             std::uniform_int_distribution<uint64_t> keys_limit_dist(0, 8192);
 
-            te->add_params()->set_num(keys_limit_dist(rg.generator));
+            te->add_params()->set_num(static_cast<uint32_t>(keys_limit_dist(rg.generator)));
         }
     }
     else if (te->has_engine() && (b.isAnyIcebergEngine() || b.isAnyDeltaLakeEngine() || b.isAnyS3Engine() || b.isAnyAzureEngine()))
